% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_obs.R
\name{bin_obs}
\alias{bin_obs}
\title{Bin continuous observations into categories}
\usage{
bin_obs(x, ncats)
}
\arguments{
\item{x}{Vector of observations to bin}

\item{ncats}{Number of desired categories to output}
}
\value{
Vector of integers
}
\description{
Bin continuous observations into categories
}
\examples{
bin_obs(seq(0, 1, by = 0.1), 5)
}
