% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helper.R
\name{date_helper}
\alias{date_helper}
\title{A function to format dates and times stored in separate columns}
\usage{
date_helper(date, time, tz = NULL, date_format = "\%m/\%d/\%Y",
  time_format = "\%H:\%M")
}
\arguments{
\item{date}{A character representation of the date}

\item{time}{A character representation of the time}

\item{tz}{The time zone with which the data was collected}

\item{date_format}{An character string reflecting the storage format of the dates, to view options see ?base::strptime}

\item{time_format}{An character string reflecting the storage format of the time, to view options see ?base::strptime}
}
\value{
A vector of class POSIXct
}
\description{
A function to format dates and times stored in separate columns
}
\examples{
tst <- tibble::tibble(
  Date = "10/01/2017",
  Time = "11:33"
)

tst <- date_helper(tst$Date, tst$Time, tz = "UTC")

}
