% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yr.R
\name{get_yr}
\alias{get_yr}
\title{Create integer year to be used for index}
\usage{
get_yr(dt_col)
}
\arguments{
\item{dt_col}{Vector of dates}
}
\value{
vector of integers
}
\description{
Create integer year to be used for index
}
\examples{
tst <- tibble::tibble(
  dates = Sys.time()
)

get_yr(tst$dates)
}
