% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jul_day.R
\name{jul_day}
\alias{jul_day}
\title{Create julian day from date time column}
\usage{
jul_day(dt_col)
}
\arguments{
\item{dt_col}{Vector of dates}
}
\value{
vector of integers
}
\description{
Create julian day from date time column
}
\examples{
tst <- tibble::tibble(
  dates = Sys.time()
)

jul_day(tst$dates)
}
