% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph_data.R
\name{morph_data}
\alias{morph_data}
\title{Wrapper for initial data handling steps}
\usage{
morph_data(path, ncats = 5)
}
\arguments{
\item{path}{Path to detection data}

\item{ncats}{Number of categories to bin observations into}
}
\value{
A tibble with formatted columns
}
\description{
Wrapper for initial data handling steps
}
\examples{
morph_data("C:/TEMP/brandon/detections.csv")
}
