% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_detections.R
\name{read_detections}
\alias{read_detections}
\title{A function to read in detection data}
\usage{
read_detections(path, ctype = "icccicdi", ...)
}
\arguments{
\item{path}{Is the path to the file to read}

\item{ctype}{Column types of the data being read}

\item{...}{other arguments of rear::read_csv}
}
\value{
A tibble
}
\description{
A function to read in detection data
}
\examples{
tst <- read_detections(
"C:/TEMP/brandon/detections.csv",
col_names = c("picture", "camera", "date", "time", "image", "species",
  "white", "N_animals")
)
}
